use INTEGER insteadof TINYINT (#85)
* Validation table update

ALTER TABLE Validation MODIFY status INTEGER NOT NULL

+++ entity/src/main/resources/db/migration/V1.12__update_validation_status.sql
@@ -13,4 +13,4 @@ SET status = CASE
                  ELSE '1'
                  END;
 
-ALTER TABLE Validation MODIFY status tinyint NOT NULL;
+ALTER TABLE Validation MODIFY status INTEGER NOT NULL;
